Improvements to GMLExporter and GMLImporter
Changed GMLExporter to always use the integer provider for vertex and
edge IDs in the exported output, because examples of GML seem to always
use integers, and the GMLImporter is hard-coded to only work with
integer IDs.  Integers are still turned into strings in a separate step
before being sent to the file--this ought to be fixed.

GMLImporter was not observing labels for vertices or edges.  Labels can
contain data needed to reconstruct a graph properly.  Modified
GMLImporter to read the label strings for both vertices and edges.
Labels and IDs were being confused when passed to the EdgeProvider and
VertexProvider; corrected this so that the labels are passed and the IDs
are not.