add utility to create Maps/Sets with proper capacity for expected size
The int-argument in the constructor of java.util.(Linked)HashMap
specifies not an initial size of the Map rather than an initial capacity
of its internal storage. By default the internal storage is not fully
populated and therefore resized before the size of the map reaches its
storage capacity. Rehashing adds additional runtime when populating a
Map. The load-factor specifies the maximal allowed ratio of size to
capacity. Before the size to capacity ratio exceeds the load factor the
storage is enlarged.

Because LinkedHashMap, HashSet and LinkedHashSet extend HashMap and
provide an equivalent constructor that issue occurs for them, too.

The added utility class provides methods to consider the load-factor of
Maps/Sets and avoids rehashing if the final size is known before the
Map/Set is constructed.