Make DoublyLinkedList a fully capable java.util.List and Deque (#830)
Implement the java.util.List interface for DoublyLinkedList and enable
all of its optional operations, using AbstractSequentialList as base
implementation and implement a ListIterator that is able to add, remove
and set elements.

Additionally implement the Deque interface for DoublyLinkedList.

Make it save to remove and add ListNodes directly and added methods do
this.

Add tests.