[GraphImporter/Exporter] Close File-Stream in default implementations
The Reader respectively the Writer are created as a resource of a
try-with-resources statement, so they are closed automatically when the
try-block is left.

Replace the FileInputStream-creation wrapped in a InputStreamReader by a
single FileReader creation, which is equivalent but shorter and
corresponds to GraphExporter where a FileWriter is used.

Additionally I updated the Javadoc of the Reader/Writer- and
InputStream/OutputStream-argument methods to raise awareness that the
caller is responsible for closing the Stream or Reader/Writer.