Fix testNoOutDegreeZero
Decreasing the probability of alpha to 0.1, and allowing
a number of failures up to 0.05.
I am against fixing a succeeding seed. instead, we can accept
a small probability of failure.
Although decreasing alpha value should be enough, I allowed this
small failure rate because this is a stochastic distribution after all.