Replace Map<Integer,E> with ArrayList<E> where possible
Using a Integer-key in maps is only useful if the keys are not
consecutive values. In case of consecutive integer values using an
ArrayList is much faster (getting an element at a certain index of a
ArrayList is much faster than getting the value of a key from a HashMap
plus autoboxing from int to integer objects is avoided) and less memory
demanding (an ArrayList of the same size has a smaller backing array and
also does not need the Entry-objects for each entry) compared to using
an HashMap.

Map<Integer,Integer> are replaced by int-arrays if possible.