Revert buggy reduction of Map-lookups in IntrusiveEdgesSpecifics
The problem was that the 'source' and 'target' field of IntrusiveEdge,
which are used in hashCode() and equals() of IntrusiveEdge, are changed
after an Edge is put into the map. Changing the fields considered in
hashCode() and equals() after the instance is put into a HastMap/-Set is
a bug in general.

The optimization of IntrusiveEdgesSpecifics has to be done more
specifically. But the IntrusiveEdgesSpecifics can be improved
independently of the other classes changed in this PR, so I will do it
in another PR (if the other approach works).