Remove up-front check in addEdge(V, V, E) if edge is already contained
In most cases this check is just a performance penalty. And because the
rejection of an edge from the EdgeSpecific has to be handled anyway
since it can throw an exception there is no need to ensure up front that
adding will succeed.
An up-front check only saves runtime in case the edge is already present
in this or another graph but with other touching vertices (and is not an
intrusive edge). Compared to other cases (for example if just a valid
edge is added) it is probably rare. Also if it can be decided which case
to prefer it should be the valid one.