[CheckStyle] suppress Variable/Method/Type-name rules in API elements
Suppress all Variable/Method/Type name violations on API elements.
Fixing those rules requires API modifications.

In case of (static) fields, the fields are already deprecated and a
replacement is provided.

org.jgrapht.nio.DefaultAttribute : make static field 'NULL' final. This
is actually an API break, but I assume it was never intended to make it
modifiable and it was not-final by accident.