[TSPLibImporter] fix distance computation for edge-type GEO (#1064)
The conversion function from sexagesimal/base60 value to radian decimal
values in TSPLIBImporter computeRadiansAngle() had a bug when splitting
the sexagesimal/base60 value into its degree and minute part.

The integer degree part must be computed by truncating the decimal part
of the double value (by casting it to an int) instead of rounding it to
the next integer value. In case the value was rounded upwards (in case
of a negative value rounded downwards) the minute part (which is then
subtracted) becomes bigger than it should be.
This is caused by the non-steadiness of the conversion from decimal
degrees to sexagesimal/base60 degree values.

Updated and extended the test-cases accordingly.