Add callback api to DirectedSimpleCycles
Currently some relatively large manually created graphs end up using
an excessive amount of memory when detecting cycle, ultimately causing
out of memory Errors due to the sheer amount of detected cycles that are
held in a list.

While the HawickJamesSimpleCycles implementation already provides the
path limit as a way to mitigate this problem, there are use cases where
not all cycles matter to the application and as such this commit adds a
callback API to all DirectedSimpleCycle implementations allowing users
of this library to decide how to deal with each detected cycle, without
forcing the the whole list of detected cycles to remain in memory until
the algorithm finishes.